refactoring player and creatures

+++ src/Rasa.Game/Packets/MapChannel/Client/AssignNPCMissionPacket.cs
@@ -8,13 +8,13 @@
         public override GameOpcode Opcode { get; } = GameOpcode.AssignNPCMission;
 
         public long NpcEntityId { get; set; }
-        public int MissionId { get; set; }
+        public uint MissionId { get; set; }
 
         public override void Read(PythonReader pr)
         {
             pr.ReadTuple();
             NpcEntityId = pr.ReadLong();
-            MissionId = pr.ReadInt();
+            MissionId = (uint)pr.ReadInt();
         }
     }
 }
