+++ src/Rasa.Game/Packets/MapChannel/Client/PersonalInventory_DestroyItemPacket.cs
@@ -8,13 +8,13 @@
         public override GameOpcode Opcode { get; } = GameOpcode.PersonalInventory_DestroyItem;
 
         public long EntityId { get; set; }
-        public long Quantity { get; set; }
+        public uint Quantity { get; set; }
 
         public override void Read(PythonReader pr)
         {
             pr.ReadTuple();
             EntityId = pr.ReadLong();
-            Quantity = pr.ReadLong();
+            Quantity = (uint)pr.ReadLong();
         }
     }
 }
